+++ androidsdk/src/main/java/com/wootric/androidsdk/network/tasks/CreateResponseTask.java
@@ -1,5 +1,6 @@
 package com.wootric.androidsdk.network.tasks;
 
+import com.wootric.androidsdk.Constants;
 import com.wootric.androidsdk.OfflineDataHandler;
 
 /**
@@ -8,21 +9,28 @@ import com.wootric.androidsdk.OfflineDataHandler;
 public class CreateResponseTask extends WootricRemoteRequestTask {
 
     private final long endUserId;
+    private final long userId;
+    private final long accountId;
     private final String originUrl;
     private final int score;
+    private final int priority;
     private final String text;
+    private final String uniqueLink;
 
     private final OfflineDataHandler offlineDataHandler;
 
-    public CreateResponseTask(long endUserId, String originUrl, int score, String text,
-                              String accessToken, OfflineDataHandler offlineDataHandler) {
+    public CreateResponseTask(long endUserId, long userId, long accountId, String originUrl, int score, int priority, String text, String accessToken, OfflineDataHandler offlineDataHandler, String uniqueLink) {
         super(REQUEST_TYPE_POST, accessToken, null);
 
         this.endUserId = endUserId;
+        this.userId = userId;
+        this.accountId = accountId;
         this.originUrl = originUrl;
         this.score = score;
+        this.priority = priority;
         this.text = text;
         this.offlineDataHandler = offlineDataHandler;
+        this.uniqueLink = uniqueLink;
     }
 
     @Override
@@ -32,14 +40,26 @@ public class CreateResponseTask extends WootricRemoteRequestTask {
 
     @Override
     protected void buildParams() {
+        paramsMap.put("end_user_id", String.valueOf(endUserId));
+        if (userId != (long) Constants.NOT_SET) {
+            paramsMap.put("user_id", String.valueOf(userId));
+        }
+        paramsMap.put("priority", String.valueOf(priority));
         paramsMap.put("origin_url", originUrl);
         paramsMap.put("score", String.valueOf(score));
-        addOptionalParam("text", text);
+        paramsMap.put("survey[channel]", "mobile");
+        paramsMap.put("survey[unique_link]", uniqueLink);
+        if (accountId != (long) Constants.NOT_SET) {
+            paramsMap.put("account_id", String.valueOf(accountId));
+        }
+        if (!text.equals("")) {
+            addOptionalParam("text", text);
+        }
     }
 
     @Override
     protected void onError(Exception e) {
         super.onError(e);
-        offlineDataHandler.saveOfflineResponse(endUserId, originUrl, score, text);
+        offlineDataHandler.saveOfflineResponse(endUserId, userId, accountId, originUrl, score, priority, text, uniqueLink);
     }
 }